requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
pymupdf==1.23.7
pytesseract==0.3.10
Pillow==10.1.0
google-genai>=1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
aiofiles==23.2.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pytest==7.4.3
